requirements.txt
torch --extra-index-url https://download.pytorch.org/whl/cu128
torchvision --extra-index-url https://download.pytorch.org/whl/cu128
torchaudio --extra-index-url https://download.pytorch.org/whl/cu128
matplotlib
numpy
streamlit
Pillow